Methoden ausgelagert
