+++ Assets/Scripts/Ball/Ball.cs
@@ -7,7 +7,7 @@ using UnityEngine.Serialization;
 using Debug = UnityEngine.Debug;
 using Random = UnityEngine.Random;
 
-public class Ball : MonoBehaviour
+public class Ball : Resettable
 {
     [SerializeField] [Range(10, 80)] float randomSpeed = 40;
     [SerializeField] float initialForce;
@@ -19,14 +19,6 @@ public class Ball : MonoBehaviour
     private const float MINVelocity = 0.4f;
     private const float MINAngularVelocity = 1.047f;
     private float _lastStoppedTime;
-
-
-    private Vector3 _position;
-    private Vector3 _velocity;
-    private Quaternion _rotation;
-    private Vector3 _angularVelocity;
-    private Rigidbody _rb;
-    private Transform _transform;
     private const float TimeWindowToStop = 2.0f;
 
     void Start()
@@ -52,30 +44,6 @@ public class Ball : MonoBehaviour
         if (Input.GetButtonDown("Select"))
             ResetShot(new Vector3(7.76f, 2.98f, 0f));
         _rb.velocity = Vector3.ClampMagnitude(_rb.velocity, maxVelocity);
-    }
-
-    private void FixedUpdate()
-    {
-        Clone( _rb);
-    }
-
-    private void Clone(Rigidbody rb)
-    {
-        
-        _position = rb.position;
-
-        _position = new Vector3
-        {
-            x = _position.x,
-            y = _position.y,
-            z = _position.z,
-        };
-        _rotation = new Quaternion((_rotation = rb.rotation).x, _rotation.y, _rotation.z, _rotation.w);
-        _angularVelocity = new Vector3(
-            _angularVelocity.x,
-            _angularVelocity.y,
-            _angularVelocity.z);
-        _velocity = new Vector3(_velocity.x, _velocity.y, _velocity.z);
     }
 
     private void LateUpdate()
@@ -165,41 +133,13 @@ public class Ball : MonoBehaviour
     {
         if (!col.gameObject.CompareTag("Ground"))
         {
-            SetBallOneFrameBack();
-           // col.rigidbody.AddForceAtPosition(-col.impulse,col.contacts.First().point, ForceMode.Impulse);
-            
-            // _rb.AddForceAtPosition(new Vector3(1,1,0) * 10,col.contacts.First().point, ForceMode.VelocityChange);
-            _rb.AddForce(new Vector3(1,1,0) * 10, ForceMode.VelocityChange);
-            
-            Debug.Log($" Force {col.contacts.First().point}");
-            
-            Debug.Log($" Neu : {_rb.position.x} Alt: {_position.x}");
-            Debug.Log($"Hit : {col.GetContact(0).normal} {(_rb.position - col.GetContact(0).point).normalized}");
-            //float impulseMagnitude = initialForce + col.impulse.magnitude * hitMultiplier;
-            //Vector3 dir = transform.position - col.contacts[0].point;
-            // var dir = _rb.position - col.contacts[0].point;
-            // Debug.Log($"{dir.x / dir.y}");
-            // _rb.AddForce(col.impulse, ForceMode.Impulse);
-            // _rb.AddForce(CalculatePsyonixImpulse(col), ForceMode.Impulse);
-            // _rb.AddForce(-col.impulse * hitMultiplier, ForceMode.Impulse);
-            Debug.Log($" Force : {col.impulse * hitMultiplier}");
+            CancelUnityImpulse();
+            J_psyonix = CustomPhysics.CalculateBulletImpulse(col);
+            J_bullet = CustomPhysics.CalculatePsyonixImpulse(_rb, col);
+            //TODO: Add bullet impulse to car
+            Vector3 J = J_bullet + J_psyonix;
+            _rb.AddForceAtPosition(J,col.contacts.First().point, ForceMode.VelocityChange);
         }
-    }
-
-    private Vector3 CalculateBulletImpulse()
-    {
-        return Vector3.zero;
-        
-    }
-
-    private void SetBallOneFrameBack()
-    {
-        _rb.position = _position;
-        // _rb.transform.position = _position + _rb.velocity* Time.deltaTime;
-        // _rb.transform.rotation = _rotation;
-        _rb.velocity = _velocity;
-         _rb.angularVelocity = _angularVelocity;
-        _rb.rotation = _rotation;
     }
 
     private void OnCollisionExit(Collision other)
@@ -210,27 +150,5 @@ public class Ball : MonoBehaviour
         }
     }
 
-    Vector3 CalculatePsyonixImpulse(Collision col)
-    {
-        var n = _rb.position - col.rigidbody.position;
-        n.y *= 0.35f;
-        
-        Debug.Log($"y : {n.y/n.x}");
-        var f = col.transform.forward;
-        var dot = Vector3.Dot(n, f);
-        n = Vector3.Normalize(n - 0.35f * dot * f);
-        var impulse = _rb.mass * Math.Abs(col.relativeVelocity.magnitude) * scaling(col.relativeVelocity.magnitude) *
-                      n;
-        //Debug.Log(impulse);
-        //Debug.Log(col.relativeVelocity.magnitude);
-        return impulse;
-    }
-
-    float scaling(float magninute)
-    {
-        var test = pysionixImpulseCurve.Evaluate(magninute/10f);
-        
-        Debug.Log($"{Time.time}: Curve {test} : Value {magninute}");
-        return test;
-    }
+    
 }
