+++ Assets/Scripts/CarControllers/CarCollision.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 
-public class CarCollision : MonoBehaviour
+public class CarCollision : Resettable
 {
     private MatchController.MatchController _matchController;
     private float forwardSpeed;
@@ -21,6 +21,7 @@ public class CarCollision : MonoBehaviour
 
     private void FixedUpdate()
     {
+        //TODO: Conform to Resettable
         forwardSpeed = Vector3.Dot(GetComponent<Rigidbody>().velocity, transform.forward);
     }
