+++ Assets/Scripts/Resettable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 094ec2868eada3b428711552b9c80325
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
